requirements.txt
asgiref==3.4.1
celery==5.1.2
Django==3.2.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.8.1
djangorestframework==3.12.4
pre-commit==2.15.0
psycopg2==2.9.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
sqlparse==0.4.2
